BaddieState physicsActive, moveJustStartedB and stateTag are s8
Retail load census over the whole DOL, keyed on the struct offset:

  0x25F physicsActive     9 loads, 9 lbz+extsb, 0 zero-extended
  0x27B moveJustStartedB  125 loads, 118 lbz+extsb, 0 zero-extended
                          (the 7 remaining are TumbleweedState/MagicDust,
                          a different struct at the same offset)
  0x34D stateTag          6 loads, 6 lbz+extsb, 0 zero-extended

No reader anywhere takes the zero-extended value: there is no cmplwi, no
andi, no clrlwi and no widening pass of any of the three. Retype them s8
in BaddieState and in the three structs that alias the physicsActive slot
(TrickyState, TreasureChestState, KtRexState), then drop the 37 reader
casts and address puns that were spelling the signedness by hand.

Every object byte-identical; main.dol OK.

+++ src/dlls/objects/203/203.c
@@ -75,7 +75,7 @@ int dll_CB_stateHandler5(GameObject* obj, GroundBaddieState* state) {
 }
 
 int dll_CB_stateHandler4(GameObject* obj, GroundBaddieState* state) {
-    if ((s8)state->baddie.moveJustStartedB != 0) {
+    if (state->baddie.moveJustStartedB != 0) {
         f32 initialVelocity;
 
         (*gPlayerInterface)->setState(obj, state, 0);
@@ -105,7 +105,7 @@ int dll_CB_stateHandler4(GameObject* obj, GroundBaddieState* state) {
 int dll_CB_stateHandler3(GameObject* obj, GroundBaddieState* state) {
     GroundBaddieState* objectState = obj->extra;
 
-    if ((s8)state->baddie.moveJustStartedB != 0) {
+    if (state->baddie.moveJustStartedB != 0) {
         (*gBaddieControlInterface)->spawnChild(obj, objectState->triggerId, -1, 0);
     }
     return 0;
@@ -114,7 +114,7 @@ int dll_CB_stateHandler3(GameObject* obj, GroundBaddieState* state) {
 int dll_CB_stateHandler2(GameObject* obj, GroundBaddieState* state) {
     ObjHitsPriorityState* hitState;
 
-    if ((s8)state->baddie.moveJustStartedB != 0) {
+    if (state->baddie.moveJustStartedB != 0) {
         (*gPlayerInterface)->setState(obj, state, 3);
         state->baddie.targetObj = NULL;
         state->baddie.physicsActive = 0;
